+++ angular-workspace/projects/example-client-app/src/app/login/login.component.ts
@@ -9,10 +9,4 @@ import { NimbleTheme } from '@ni/nimble-components/dist/esm/theme-provider/theme
 export class LoginComponent {
     public theme: NimbleTheme = NimbleTheme.Light;
     public themes = NimbleTheme;
-    private toggle = true;
-
-    public switchThemes(): void {
-        this.toggle = !this.toggle;
-        this.theme = this.toggle ? NimbleTheme.Light : NimbleTheme.Dark;
-    }
 }