+++ angular-workspace/projects/ni/nimble-angular/src/directives/listbox-option/index.ts
@@ -1,0 +1,2 @@
+export * from './nimble-listbox-option.module';
+export * from './nimble-listbox-option.directive';