+++ packages/nimble-components/src/tests/utilities/theme-test-helpers.ts
@@ -41,6 +41,12 @@ export const invalidStates = [
 ];
 export type InvalidState = typeof invalidStates[number];
 
+export const readOnlyStates = [
+    ['', ''],
+    ['Read-Only', 'readonly']
+];
+export type ReadOnlyState = typeof readOnlyStates[number];
+
 /**
  * Wraps a given component template with a region for each of the available themes.
  * Takes an optional array of state values that can be used with the template to match the permutations of the provided states.
@@ -72,6 +78,17 @@ export function matrixThemeWrapper<State1, State2, State3, State4>(
     dimensions: [State1[], State2[], State3[], State4[]]
 ): string;
 
+export function matrixThemeWrapper<State1, State2, State3, State4, State5>(
+    component: (
+        state1: State1,
+        state2: State2,
+        state3: State3,
+        state4: State4,
+        state5: State5
+    ) => string,
+    dimensions: [State1[], State2[], State3[], State4[], State5[]]
+): string;
+
 export function matrixThemeWrapper(
     component: (...states: unknown[]) => string,
     dimensions?: unknown[][]
@@ -79,7 +96,7 @@ export function matrixThemeWrapper(
     const matrix: string[] = [];
     const recurseDimensions = (
         currentDimensions?: unknown[][],
-        ...states
+        ...states: unknown[]
     ): void => {
         if (currentDimensions && currentDimensions.length >= 1) {
             const [currentDimension, ...remainingDimensions] = currentDimensions;